% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstabulation.R
\name{crosstab}
\alias{crosstab}
\title{Crosstabulation}
\usage{
crosstab(data = ., x, y)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{x}{Name of the first variable. Needs to be entered as string
character.}

\item{y}{Name of the second variable. Needs to be entered as string
character.}
}
\value{
A crosstab of two variables.
}
\description{
A quick function to get a crosstabulation of two variables in a dataframe.
}
\examples{
crosstab(data, "gender", "gpa")

}
