% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics_freq_table.R
\name{demographic_table}
\alias{demographic_table}
\title{DEMOGRAPHICS FREQUENCY TABLE}
\usage{
demographic_table(data = ., variables)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{variables}{A required vector of column names representing their 
manifest demographic variables.}
}
\value{
A dataframe of the frequencies of each category
}
\description{
Create a demographic frequency table that provides the count and
proportion of each variable value. The current version does not support
automatically extracting variable labels from SPSS files through \code{haven}
(Wickham et al., 2023).
The variable values will be returned as is, meaning that if the variables
are numeric, then the table will display numeric values. Users are
recommended to recode the demographic variables of interest with string
values as necessary.
}
\examples{
variables <- c("age", "education", "gender")
demographic_table(data, variables)

}
\references{
Wickham H, Miller E, Smith D (2023). _haven: Import and Export
'SPSS', 'Stata' and 'SAS' Files_. R package version 2.5.2,
\url{https://CRAN.R-project.org/package=haven}.
}
