% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_table.R
\name{freq}
\alias{freq}
\title{Frequency Table}
\usage{
freq(data = ., variable)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{variable}{A required single string denoting the name of the variable.}
}
\value{
A frequency table of the variable's unique values.
}
\description{
A quick function to get a frequency table of a variable's values.
}
\examples{
freq(data, "education")

}
