% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_least_squares_sem_wrapper.R
\name{plssem_wrapper}
\alias{plssem_wrapper}
\title{Partial Least Squares Structural Equation Modeling Wrapper}
\usage{
plssem_wrapper(data = ., measurements, structure, file = NULL, bootn = 1000)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{measurements}{Measurement model object from \code{seminr} package. Refer
to the \code{seminr} help documentation for more details.}

\item{structure}{Structural model object from \code{seminr} package.}

\item{file}{Location to save output file as excel workbook if specified.}

\item{bootn}{Number of bootstrap replications to calculate p-values at the
structural pathways. Default to 1000.}
}
\description{
A wrapper function to automate PLS-SEM and extracting its results. The
current function is limited to only 2 serial mediators. Due to the undue
level of complexity added to the model estimations when there are 3 or
more serial mediators, the current function is not designed to handle this
type of model. This automation function uses the \code{seminr} package (Ray et
al., 2022).
}
\examples{
measurements <- constructs(
  composite("comp1", multi_items("com", 1:3)),
  composite("comp2", multi_items("com", 4:6)),
  higher_composite("high_comp", c("comp1", "comp2")),
  composite("mani1", c("v1")),
  composite("mani2", c("v2")),
  composite("outcome", multi_items("out", 1:5)),
  interaction_term(iv = "high_comp", moderator = "mani1", method = two_stage)
)

structure <- relationships(
  paths(from = c("high_comp"),
        to = c("mani1", "mani2")),
  paths(from = c("high_comp", "mani1", "mani2", "high_comp*mani1"),
        to = c("outcome"))
)

plssem_wrapper(data,
               measurements,
               structure)


}
\references{
Ray S, Danks N, Calero Valdez A (2022). seminr: Building and
Estimating Structural Equation Models. R package version 2.3.2,
\url{https://CRAN.R-project.org/package=seminr}.
}
