% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions_test_wrapper.R
\name{proportion_test}
\alias{proportion_test}
\title{Proportions Test Wrapper}
\usage{
proportion_test(data = ., group, outcomes)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{group}{A single character string of the group variable name. The
variable should not be factored with different levels. The current wrapper
does not support this capability.}

\item{outcomes}{Vector of outcome variable names. The outcomes should have
the same levels (e.g., all 0's and 1's, or all 1's and 2's).}
}
\value{
A list of dataframes containing the chi-square proportions test
results, the proportions of each group's levels, and the pairwise comparison
results.
}
\description{
A simple wrapper that automates the running and extracting of common report
metrics for a chi-square independent proportions test.
Automatically applies holm p adjustment to multiple comparisons. Returns
model statistic and also pairwise comparison results.
}
\examples{
data <- carData::WVS \%>\% 
   mutate(across(.cols = c(poverty, religion, degree, gender),
                 .fns = function(x) as.integer(x)),
          country = as.character(country))

proportion_test(data = data,
                group = "country",
                outcomes = c("degree",
                             "gender"))

}
