% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_equation_model_wrapper.R
\name{sem_wrapper}
\alias{sem_wrapper}
\title{Structural Equation Modeling Wrapper}
\usage{
sem_wrapper(
  data = .,
  model,
  cluster = NULL,
  predictors = NULL,
  estimator = "ML",
  se = NULL,
  missing = "listwise"
)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{model}{`Lavaan` defined SEM object. Refer to \code{lavaan}
documentation for a description of the user-specified model.}

\item{cluster}{If working with clustered/multilevel/hierarchical data, define
the cluster or group variable to account for shared SE. Cluster should be
a categorical factor variable.}

\item{predictors}{An optional vector of character strings of the predictors
of interest. Doing so will filter the coefficient table to just the
predictors of interest for easier viewing.}

\item{estimator}{Estimator to be used in the confirmatory factor analysis.
Default is maximum likelihood (ML). Alternative esimators are limited to
those offered by \code{lavaan}, e.g., "GLS", "WLS", "DWLS", "ULS", "DLS", and
"PML". Refer to \code{lavaan} documentation for additional details.}

\item{se}{Parameter to determine how to compute standard errors. Default
is set to NULL. Set to "robust" to use either "robust.sem" or
"robust.huber.white" depending on the estimator used. Refer to lavaan
documentation for additional details.}

\item{missing}{Parameter to determine how to handle missingness. Default
is set to "listwise" where all observations with missing values are deleted
prior to analysis. As detailed in \code{lavaan}'s documentation, this is only
recommended when data are missing completely at random (MCAR). Another
options is "ml" for full information maximum likelihood approach (fiml).
Refer to lavaan documentation for additional details.}
}
\value{
A list of two additional lists. The first list consists of the
standard \code{lavaan} SEM object. The second list is a series of dataframes
containing common outputs of an SEM object compiled into an exportable
format to an excel document.
}
\description{
A simple wrapper that automates the running and extraction of common
SEM fit metrics, coefficients, and fit metrics.
}
\examples{
# Note: Model example courtesy of 'lavaan'

model <- '
  # measurement model
    ind60 =~ x1 + x2 + x3
    dem60 =~ y1 + y2 + y3 + y4
    dem65 =~ y5 + y6 + y7 + y8
  # regressions
    dem60 ~ ind60
    dem65 ~ ind60 + dem60
  # residual correlations
    y1 ~~ y5
    y2 ~~ y4 + y6
    y3 ~~ y7
    y4 ~~ y8
    y6 ~~ y8
 '

sem_wrapper(data = data,
            model = model)

}
\references{
Yves Rosseel (2012). lavaan: An R Package for Structural Equation
Modeling. \emph{Journal of Statistical Software}, 48(2), 1-36.
https://doi.org/10.18637/jss.v048.i02
}
