% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_train_test_splitter.R
\name{test_split}
\alias{test_split}
\title{Split Data Into Train, Validation, and Test Sets}
\usage{
test_split(data = ., group = NULL, proportion = 0.6, validation = FALSE)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{group}{If there are subgroups in the dataset that also need to be
equally split into training, validation, or test sets, specify the column
name reflecting the group as a character. Default is NULL}

\item{validation}{Default is FALSE. If there is a reason to have a
validation set in addition to a test set, specify this as TRUE. The
remaining data left from partitioning the training set are then randomly
and equally split into the validation and test sets.}

\item{fraction}{Percentage of the dataset that should be subset as the
training set. Default is 0.6 to indicate that 60% of the dataset will be
randomly sampled to comprise the training set.}
}
\value{
A list of 2 or 3 dataframes representing the training, validation,
and test sets.
}
\description{
Automated wrapper function to split a dataframe into training, validation,
and testing sets for machine learning or cross-validation. Validation and
test sets are equally split from the remaining dataframe after subsetting
the training set. Training, validation, and test sets are saved as
dataframes in a returned list.
}
